Introduce function_id for instrumented functions
Previously the client code used absolute function address to identify
the instrumented function in timers and in some other places in UI.

Since using the address assumes that the function address in memory is
fixed (which might not be true) this has been replaced with function_id
for instrumented functions.

Changed code relying on function absolute address where it is not needed
and replaced it with function_id.

Test: run unit tests
Test: hook functions run capture, create iterators and frame tracks
Bug: http://b/177525225
Bug: http://b/175865913

+++ src/OrbitGl/DataView.h
@@ -91,7 +91,7 @@ class DataView {
   virtual void DoFilter() {}
   FilterCallback filter_callback_;
 
-  std::vector<uint32_t> indices_;
+  std::vector<uint64_t> indices_;
   std::vector<SortingOrder> sorting_orders_;
   int sorting_column_ = 0;
   std::string filter_;
