+++ src/OrbitGl/SamplingReportDataView.cpp
@@ -301,7 +301,7 @@ void SamplingReportDataView::SetThreadID(ThreadID tid) {
 }
 
 void SamplingReportDataView::DoFilter() {
-  std::vector<uint32_t> indices;
+  std::vector<uint64_t> indices;
 
   std::vector<std::string> tokens = absl::StrSplit(ToLower(filter_), ' ');
