+++ src/OrbitGl/TracepointsDataView.cpp
@@ -77,7 +77,7 @@ void TracepointsDataView::DoSort() {
 }
 
 void TracepointsDataView::DoFilter() {
-  std::vector<uint32_t> indices;
+  std::vector<uint64_t> indices;
   std::vector<std::string> tokens = absl::StrSplit(ToLower(filter_), ' ');
 
   for (size_t i = 0; i < tracepoints_.size(); ++i) {
